ui/requirements.txt
streamlit==1.28.1
pandas==2.1.4
pillow==10.1.0
